Move result classes to separate package
